#  Faster VCFs loading, support HTTP, and refactored variant metadata