--- a/test/data/multiallelic.vcf
+++ b/test/data/multiallelic.vcf
@@ -12,5 +12,5 @@
 ##contig=<ID=chr15,length=102531392>
 ##contig=<ID=chr16,length=90354753>
 ##contig=<ID=chr17,length=81195210>
-#CHROM  POS ID  REF ALT QUAL    FILTER  INFO    FORMAT  metastasis
-chr1    1431105 rs199599542     A       C,G     593.69  PASS DP=17;GE=Wuzzle     GT     0/1
+#CHROM  POS ID  REF ALT QUAL	FILTER  INFO	FORMAT  metastasis
+chr1	1431105	rs199599542	A	C,G	593.69	PASS	DP=17;GE=Wuzzle	GT	0/1