--- a/varcode/vcf.py
+++ b/varcode/vcf.py
@@ -141,6 +141,9 @@ def load_vcf_fast(
 
     This is an experimental faster implementation of `load_vcf`. It is
     typically about 2X faster, and with `include_info=False`, about 4X faster.
+    If most of the records in the VCF have failed filters (and
+    only_passing=True), this function can be orders of magnitude faster than
+    `load_vcf`.
 
     Currently only local files are supported by this function (no http). If you
     call this on an HTTP URL, it will fall back to `load_vcf`.
@@ -472,4 +475,3 @@ def parse_url_or_path(s):
     if s.startswith("//"):
         s = "/" + s.lstrip("/")
     return urlparse(s)
-