… test/data/mutect-example-headerless.vcf
… test/data/mutect-example.vcf
… test/data/somatic_hg19_14muts.vcf.gz
… test/data/strelka-example.vcf
… test/test_maf.py
… test/test_variant.py
… test/test_variant_collection.py
… test/test_vcf.py