There is a small related pyensembl PR, which I'll push shortly. 

 * Speed improvements to `load_vcf`, using a new pandas-based implementation and other tweaks. This is 2X faster if you need the variant info column, and 9X faster if you don't.
 * Add support for loading VCFs over any protocol supported by the `requests` package (HTTP, HTTPS, FTP). Closes #91.
 * Refactoring: got rid of the `variant.info` field and moved it to `VariantCollection.metadata`, which is a dict from variants to extra metadata. I think this simplifies things, especially when working with variants coming from different sources, where the variant may be same but the metadata may be different.

Optimizations
 * Specify a default sort key in VariantCollection.__init__, to avoid making variant pairwise comparisons in sorted() in Collection.__init__.
 * Optimize memoize function for common case (cache hit).
 * Avoid unnecessary list creation in Collection.__init__
 * Faster implementation of trim_shared_suffix
 * Use slots for Variant instances.
 * Optimize common case of a simple SNV in Variant.__init__
 * Optimize Variant.__hash__, __eq__

Other changes
 * Add test/benchmark_vcf_load.py for benchmarking `load_vcf` performance.
 * Add allow_extended_nucleotides parameter to load_vcf. This is necessary to load the dbsnp VCF.
 * Change multiallelic.vcf test case to use tabs (as per VCF standard) not spaces.

Timing details. Time to load first 100K variants in dbsnp VCF (median of 3 runs):
 * master varcode + pyensembl: 10.8 sec
 * varcode tweaks: 8.1 sec
 * varcode tweaks + pandas-based loader: 5.4 sec
 * varcode tweaks + pandas + skip parsing info:  1.5 sec
 * varcode tweaks + pandas + skip parsing info + pyensembl tweaks:  1.2 sec

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/hammerlab/varcode/94)
<!-- Reviewable:end -->