@iskandr 2015-06-22
Small initial comment: can we make the fast path for parsing optional and PyVCF the default?