@timodonnell 2015-06-24
Ok, pushed an update with this change, as well as some hopefully clearer code in vcf.py