@timodonnell 2015-06-25
Also minor note: while `load_vcf_fast` is only 2-4X faster than `load_vcf` on typical data, on files like Mutect output, where the vast majority of variants are REJECT, this is many orders of magnitude faster.