@timodonnell 2015-07-06
Updated to address the single v double quotes comment